# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aloware Test Project",
        "description": "Aloware Test Project",
        "contact": {
            "email": "nolie.francisco03@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/v1/comment": {
            "get": {
                "tags": [
                    "Fetch Comments"
                ],
                "summary": "Get list of comments",
                "operationId": "GetComment",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Post Comment"
                ],
                "summary": "Store/Add comment",
                "operationId": "PostComment",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "comment",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Page Not found"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        }
    }
}